Move simulation functions to HLAfreq
These functions will be used for multiple tests so should be available
